+++ NEWS.md
@@ -67,7 +67,7 @@ whether all letters are upper case and issue informative warnings and error mess
 * as.TermDocumentMatrix,partition_bundle optimized to work efficiently with large corpora.
 * Fixed an issue for `as.DocumentTermMatrix` that started to occur with the introduction of the `subcorpus_bundle` class (#100).
 * Removed a bug in the kwic,character-method that prevented using different values for right and left context (#101).
-
+* Removed a bug from the kwic,character-method, and the context,corpus-method that would result in odd behavior when either the left or right context is 0.